+++ PeteBJJCompFeed/PeteBJJCompFeediOSTests/CompetitionsViewControllerTests.swift
@@ -421,19 +421,6 @@ private extension UIButton {
     }
 }
 
-private extension UIImage {
-    static func make(withColor color: UIColor) -> UIImage {
-        let rect = CGRect(x: 0, y: 0, width: 1, height: 1)
-        let format = UIGraphicsImageRendererFormat()
-        format.scale = 1
-        
-        return UIGraphicsImageRenderer(size: rect.size, format: format).image { rendererContext in
-            color.setFill()
-            rendererContext.fill(rect)
-        }
-    }
-}
-
 private extension UIRefreshControl {
     func simulatePullToRefresh() {
         allTargets.forEach { target in
